Python wrapper and mapreduce test fixes.
Add channels to terasort test on YT and make it more native.
Fix reading of huge records in python wrapper.
Fix key columns in sort operation in python wrapper.
Fix chunksize in mutual information test.
Add chmod +x in python wrapper test.

+++ scripts/mapreduce_test/prepare.sh
@@ -1,14 +1,37 @@
 #!/bin/sh -eu
 
-echo " #!/bin/sh
-./gen_terasort 2500000" > run.sh
-chmod +x run.sh
+JOBCOUNT=1000
+JOB_RECORDS=10000000
 
-rm -f input
-touch input
-for (( i = START ; i < START + 4000; i++ ))
-do
-    echo -e "$i\t\t" >> input
-done
+if [ "$SYSTEM" = "mapreduce" ]; then
+    rm -f input
+    touch input
+    for (( i = START ; i < START + $JOBCOUNT; i++ ))
+    do
+        echo -e "$i\t" >> input
+    done
 
-$MAPREDUCE -server $SERVER -write "$INPUT" <input
+    $MAPREDUCE -server $SERVER -write "$INPUT" <input
+    $MAPREDUCE -server $SERVER -map "./gen_terasort $JOB_RECORDS $SYSTEM" -file gen_terasort \
+        -src "$INPUT" -dst "$OUTPUT" -jobcount $JOBCOUNT -threadcount 16 -opt cpu.intensive.mode=1
+
+elif [ "$SYSTEM" = "yt" ]; then
+    echo -e "
+import config
+import yt
+config.DEFAULT_PROXY='$SERVER'
+config.DEFAULT_FORMAT=yt.DsvFormat()
+
+input = '//home/ignat/' + '$INPUT'
+output = '//home/ignat/' + '$OUTPUT'
+yt.write_table(input, ['k=%d\\\n' % i for i in xrange($JOBCOUNT)])
+yt.create_table(output)
+yt.set_attribute(output, 'channels', '[[\\\"k\\\", \\\"v\\\"]]')
+spec = {'job_count': $JOBCOUNT,
+        'locality_timeout': 0}
+yt.run_map('./gen_terasort $JOB_RECORDS $SYSTEM', input, yt.Table(output, append=True), files='gen_terasort', spec=spec)
+
+" >gen.py
+    python gen.py
+
+fi
